requirements.txt
requests
pytorch
tqdm
psutil
py-cpuinfo
